% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_county.R
\name{aggregate_county}
\alias{aggregate_county}
\title{Aggregate Xwalk to County}
\usage{
aggregate_county(data)
}
\arguments{
\item{data}{Crosswalk tibble}
}
\value{
A tibble with:
\itemize{
  \item fips: county code
  \item county: county name
  \item state: state name
  \item cbsa: core based statistical area code
  \item cbsaname: core based statistical area name
 }
}
\description{
Aggegates Crosswalk file by fips or county code.
}
