% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_flow.R
\name{aggregate_flow}
\alias{aggregate_flow}
\title{Aggregate Origin Destination Flow}
\usage{
aggregate_flow(data)
}
\arguments{
\item{data}{The output from get commuter flow}

\item{level}{Specified aggregation level.}
}
\value{
A tibble
}
\description{
Aggregate Origin Destination Flow
}
