% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area_characteristics.R
\name{get_area_characteristics}
\alias{get_area_characteristics}
\title{get_area_characteristics}
\usage{
get_area_characteristics(st, year, dataset, seg = "S000", type = "JT00")
}
\arguments{
\item{st}{2-digit state abreviation}

\item{year}{Year of job data. Can have the value of 2002-2015 for most states.}

\item{dataset}{Either "rac" for residence or "wac" for workplace}

\item{seg}{Segment of the workforce, can have the values of “S000”, “SA01”, “SA02”, “SA03”, “SE01”, “SE02”, “SE03”, “SI01”, “SI02”, or “SI03”. These correspond to the same segments of the workforce as are listed in the OD file structure above.}

\item{type}{Job Type, can have a value of “JT00” for All Jobs, “JT01” for Primary Jobs, “JT02” for All Private Jobs, “JT03” for Private Primary Jobs, “JT04” for All Federal Jobs, or “JT05” for Federal Primary Jobs.}
}
\value{
A tibble
}
\description{
This loads either the Residence Area Characteristics or Workplace Area Characteristics data files. This is count data for jobs dissagregated down to census block.
}
\examples{
nj_residence <- get_area_characteristics(st = "nj", year = 2017, dataset = "rac")
}
