% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weighted_centers.R
\name{get_weighted_centers}
\alias{get_weighted_centers}
\title{get_weighted_centers}
\usage{
get_weighted_centers(st, year, rac = TRUE)
}
\arguments{
\item{st}{2-digit state abreviation}

\item{year}{year}

\item{rac}{TRUE (default) for residence center or FALSE for workplace center}
}
\value{
A tibble with fips, latitude, and longitude collumns
}
\description{
This function calculates the center of the population within the county. Collects
block code information in LODES data files and calulates geographic center. This is usefull to improve directions API calls.
}
\examples{
nj_cty_rac_centers <- get_weighted_centers("nj", 2017)
}
