% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/industry_codes.R
\docType{data}
\name{industry_codes}
\alias{industry_codes}
\title{QCEW Industry Codes and Titles (For NAICS Coded Data)}
\format{A data frame with 2,497 rows and 2 variables:
\describe{
  \item{code}{Industry Code}
  \item{industry_title}{Sector/Industry title}
  ...
}}
\source{
\url{https://data.bls.gov/cew/doc/titles/industry/industry_titles.htm}
}
\usage{
industry_codes
}
\description{
The lookup table shows the industry code and title relationships applicable to NAICS
coded QCEW data. The QCEW program uses the North American Industry Classification
System (NAICS) as to assign establishments to industries and to report industry
data at highly detailed as well as at aggregated levels. Detailed industry NAICS
coded data are available for the period from 1990 forward. For reference periods
from 1990 to 2006 they are coded to the 2002 version of NAICS. For periods from
2007 to 2011 they are coded to the 2007 version of NAICS. For periods from 2012
to 2016 they are coded to the 2012 version of NAICS, and for 2017 forward to the
2017 version.
}
\keyword{datasets}
