% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_cty_contig.R
\name{st_cty_contig}
\alias{st_cty_contig}
\title{st_cty_contig}
\usage{
st_cty_contig(st)
}
\arguments{
\item{st}{2 digit state abreviation}
}
\value{
A neighbor list with fips_o, fips_d,
}
\description{
Gets neighbor list from state counties
}
\examples{
ny_contig <- st_cty_contig("ny")
}
