% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_xwalk.R
\name{state_xwalk}
\alias{state_xwalk}
\title{Load State Crosswalk}
\usage{
state_xwalk(state)
}
\arguments{
\item{state}{Two letter state abreviation code}
}
\value{
data frame of the state meta data
}
\description{
This loads the geography crosswalk for the state file. This a relationship file
that establishes the hierarchical connection between each 2010 census tabulation
block and all higher level geographic entities supported by the OnTheMap system.
}
\examples{
new_york_info <- state_xwalk('ny')
}
